Used double (no float). Refactored Preprocess.
Double has more precision, which is needed. Preprocess now is separated
in three (more in the future) steps: create weka instances, apply
filters to these instances, create grid from them. Added more config
options

+++ src/it/smart/datamining/Point.java
@@ -7,14 +7,14 @@ package it.smart.datamining;
  *
  */
 public class Point {
-	private float latitude;
-	private float longitude;
+	private double latitude;
+	private double longitude;
 	
 	public Point() {
 		this(0, 0);
 	}
 	
-	public Point(float latitude, float longitude) {
+	public Point(double latitude, double longitude) {
 		super();
 		this.latitude = latitude;
 		this.longitude = longitude;
@@ -24,15 +24,15 @@ public class Point {
 		return latitude;
 	}
 
-	public void setLatitude(float latitude) {
+	public void setLatitude(double latitude) {
 		this.latitude = latitude;
 	}
 
-	public float getLongitude() {
+	public double getLongitude() {
 		return longitude;
 	}
 
-	public void setLongitude(float longitude) {
+	public void setLongitude(double longitude) {
 		this.longitude = longitude;
 	}
